♻️ Refactor: props 코드 위치 수정 (#45)
