+++ src/components/common/avatar/index.ts
@@ -1,2 +1,1 @@
 export { default as Avatar } from '@/components/common/avatar/Avatar'
-export type { AvatarProps } from '@/components/common/avatar/avatar.types'
