+++ src/components/common/image-upload/index.ts
@@ -1,3 +1,2 @@
-export * from './ImageUpload';
-export * from './imageUpload.types';
-export * from './imageUpload.utils';
+export * from './ImageUpload'
+export * from './imageUpload.utils'
