+++ src/components/common/markdown-editor/MarkdownPreview.tsx
@@ -1,9 +1,14 @@
 import MDEditor from '@uiw/react-md-editor'
 import remarkBreaks from 'remark-breaks'
 
-import type { MarkdownPreviewProps } from '@/components/common/markdown-editor/markdownEditor.types'
+import type { MarkdownValue } from '@/components/common/markdown-editor/markdownEditor.types'
 import '@/components/common/markdown-editor/markdown-editor.css'
 import { cn } from '@/utils/cn'
+
+export interface MarkdownPreviewProps {
+  value: MarkdownValue
+  className?: string
+}
 
 export default function MarkdownPreview({
   value,
