+++ src/components/common/markdown-editor/markdownEditor.types.ts
@@ -6,24 +6,8 @@ import type {
   TextAreaTextApi,
 } from '@uiw/react-md-editor'
 
-export type MarkdownValue = string
-
 export type PreviewMode = 'edit' | 'preview'
 
-export type HeadingLevel = 1 | 2 | 3 | 4 | 5 | 6
-
-export interface MarkdownEditorProps {
-  value: MarkdownValue
-  onChange: (newValue: MarkdownValue) => void
-  ariaLabel?: string
-  className?: string
-  placeholder?: string
-  disabled?: boolean
-}
-
-export interface MarkdownPreviewProps {
-  value: MarkdownValue
-  className?: string
-}
+export type MarkdownValue = string
 
 export type { ICommand, ExecuteState, TextAreaTextApi, ReactElement }
