+++ src/components/common/rating/index.ts
@@ -1,11 +1,3 @@
 export { default as Rating } from '@/components/common/rating/Rating'
 export { default as ratingUtils } from '@/components/common/rating/rating.utils'
 export { default as useRatingInteractions } from '@/components/common/rating/rating.interactions'
-
-export type {
-  RatingCommon,
-  RatingInteractiveProps,
-  RatingReadOnlyProps,
-  RatingProps,
-  UseRatingInteractionsOptions,
-} from '@/components/common/rating/rating.types'
