+++ src/components/common/toast/index.ts
@@ -1,3 +1,2 @@
 export { default as Toast } from '@/components/common/toast/Toast'
 export * from '@/components/common/toast/toast.styles'
-export type * from '@/components/common/toast/toast.types'
