--- src/components/common/toast/toast.types.ts
@@ -1,8 +1,0 @@
-type ToastType = 'success' | 'warning' | 'error'
-
-export interface ToastProps {
-  type: ToastType
-  title: string
-  message: string
-  className?: string
-}
